requirements.txt
pypugjs
flask
opencv-python
python-dotenv